the end

+++ src/main/java/com/candlebe/gcoach/controller/SettingController.java
@@ -1,7 +1,7 @@
 package com.candlebe.gcoach.controller;
 
 import com.candlebe.gcoach.dto.MemberDTO;
-import com.candlebe.gcoach.entity.*;
+import com.candlebe.gcoach.entity.Member;
 import com.candlebe.gcoach.repository.*;
 import com.candlebe.gcoach.security.dto.AuthMemberDTO;
 import com.candlebe.gcoach.service.MemberService;
@@ -14,10 +14,12 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 
+import java.util.Optional;
+
 @Controller
 @Log4j2
 @RequiredArgsConstructor
-@PreAuthorize("hasRole('USER')")
+@PreAuthorize("hasAnyRole('USER', 'ADMIN')")
 public class SettingController {
 
     private final MemberRepository memberRepository;
@@ -40,15 +42,21 @@ public class SettingController {
     }
 
     @PostMapping("/deleteAccount")
-    public String postDeleteAccount(@AuthenticationPrincipal AuthMemberDTO authMemberDTO) {
-        Member member = memberRepository.findByUsername(authMemberDTO.getUsername(), authMemberDTO.isFormSocial()).orElseThrow();
-        historyRepository.deleteHistories(member);
-        likeRepository.deleteLikes(member);
-        replyRepository.deleteReplies(member);
-        diaryRepository.deleteDiaries(member);
-        member.getRoleSet().clear();
-        memberRepository.save(member);
-        memberRepository.deleteMember(member.getMid());
+    public String postDeleteAccount(@AuthenticationPrincipal AuthMemberDTO authMemberDTO, Model model) {
+        Optional<Member> members = memberRepository.findByUsername(authMemberDTO.getUsername(), authMemberDTO.isFormSocial());
+        if (members.isPresent()) {
+            Member member = members.get();
+            historyRepository.deleteHistories(member);
+            likeRepository.deleteLikes(member);
+            replyRepository.deleteReplies(member);
+            diaryRepository.deleteDiaries(member);
+            member.getRoleSet().clear();
+            memberRepository.save(member);
+            memberRepository.deleteMember(member.getMid());
+            model.addAttribute("msg", "회원탈퇴 성공");
+            return "redirect:/logout";
+        }
+        model.addAttribute("msg", "회원탈퇴 실패");
         return "redirect:/logout";
     }
 }
